pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "langgraph-mini"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph==0.2.5",
    "langchain-openai==0.1.20",
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "typing_extensions==4.12.2",
    "pydantic==2.9.2",
    "langgraph-cli[inmem]>=0.1.55",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

